# services/manager/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Manager API
  version: 1.0.0
  description: |
    OpenAPI specification for the Manager service.
servers:
  - url: http://localhost:{port}/techcv/api/v1
    description: Local development server
    variables:
      port:
        default: '8080'
tags:
  - name: Health
    description: Service health endpoints
  - name: Authentication
    description: Firebase Authentication endpoints
paths:
  /health:
    get:
      tags:
        - Health
      summary: Retrieve the API health status
      operationId: getHealth
      responses:
        '200':
          description: Health information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/firebase/register:
    post:
      tags:
        - Authentication
      summary: Register a user after Firebase authentication
      operationId: postAuthFirebaseRegister
      responses:
        '201':
          description: Registration succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid Firebase token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/firebase/login:
    post:
      tags:
        - Authentication
      summary: Login with Firebase Authentication
      operationId: postAuthFirebaseLogin
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid Firebase token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Login failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    AuthResponse:
      type: object
      required:
        - userId
        - firebaseUid
        - email
      properties:
        userId:
          type: string
          format: uuid
        firebaseUid:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
    HealthResponse:
      type: object
      required:
        - status
        - checked_at
      properties:
        status:
          type: string
        checked_at:
          type: string
          format: date-time